mac_requirements.txt
psutil==7.0.0
pillow==11.2.1
pyinstaller==6.13.0
pytesseract==0.3.13
PyQt5==5.15.11
rapidfuzz==3.13.0
numpy
pyobjc
